Restructure package and make FTS per entity
