+++ src/bikidata/__init__.py
@@ -1,0 +1,2 @@
+from .main import query, build
+from .rdf2vec import build_rdf2vec
